+++ src/experimental/org/apache/jcs/auxiliary/lateral/xmlrpc/LateralXMLRPCSender.java
@@ -306,7 +306,7 @@ public class LateralXMLRPCSender implements IXMLRPCConstants
      * @param requesterId
      * @exception IOException
      */
-    public void remove( String cacheName, Serializable key, byte requesterId )
+    public void remove( String cacheName, Serializable key, long requesterId )
         throws IOException
     {
         CacheElement ce = new CacheElement( cacheName, key, null );
@@ -367,7 +367,7 @@ public class LateralXMLRPCSender implements IXMLRPCConstants
      * @param requesterId
      * @exception IOException
      */
-    public void removeAll( String cacheName, byte requesterId )
+    public void removeAll( String cacheName, long requesterId )
         throws IOException
     {
         CacheElement ce = new CacheElement( cacheName, "ALL", null );
